+++ frontend/src/pages/Index.js
@@ -92,7 +92,8 @@ function Index() {
         <SeatDialog />
         <Overview seat_type="standard" />
         <Typography variant="caption" gutterBottom>
-          Kör tullingedk/booking {meta.version}, commit {meta.hash}
+          Kör tullingedk/booking {meta.version}, commit {meta.hash} (
+          {meta.hashDate})
         </Typography>
         <Typography
           align="center"