+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next-release
 
+* frontend/backend: Show commit date
 * frontend: Update some texts and update to React v17
 
 ## v3.3.0 (released on 2020-10-19)